% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beta_To_M.R
\name{Beta_To_M}
\alias{Beta_To_M}
\title{Conversion of beta-values to M-values}
\usage{
Beta_To_M(beta)
}
\arguments{
\item{beta}{A numeric variable, vector or matrix corresponding to a beta-value. Has to be between 0 and 1.}
}
\value{
logit transform of the input beta-values.
}
\description{
This function performs a logit transform (\eqn{M=log2(beta)-log2(beta-1)}) on beta-values as single values or in vector, matrix, or data.frame form. M-values are not defined for beta-values equal to 0 or 1. A warning message will be displayed in case of presence of 0 and 1 in the matrix.
}
\examples{
vecBeta <- runif(100,0,1)
Beta_To_M(vecBeta)
}
