% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champ.load_extended.R
\name{champ.load_extended}
\alias{champ.load_extended}
\title{Modified champ.load function from the \emph{ChAMP} package}
\usage{
champ.load_extended(
  directory = getwd(),
  method = "ChAMP",
  methValue = "B",
  autoimpute = TRUE,
  filterDetP = TRUE,
  ProbeCutoff = 0,
  SampleCutoff = 0.1,
  detPcut = 0.01,
  filterBeads = TRUE,
  beadCutoff = 0.05,
  filterNoCG = TRUE,
  filterSNPs = TRUE,
  population = NULL,
  filterMultiHit = TRUE,
  filterXY = TRUE,
  force = FALSE,
  arraytype = "450K",
  sampleSheet.csv = "csv$",
  preproc = "Noob",
  dyeMethod = "single"
)
}
\arguments{
\item{directory}{See ChAMP manual.}

\item{method}{See ChAMP manual.}

\item{methValue}{See ChAMP manual.}

\item{autoimpute}{See ChAMP manual.}

\item{filterDetP}{See ChAMP manual.}

\item{ProbeCutoff}{See ChAMP manual.}

\item{SampleCutoff}{See ChAMP manual.}

\item{detPcut}{See ChAMP manual.}

\item{filterBeads}{See ChAMP manual.}

\item{beadCutoff}{See ChAMP manual.}

\item{filterNoCG}{See ChAMP manual.}

\item{filterSNPs}{See ChAMP manual.}

\item{population}{See ChAMP manual.}

\item{filterMultiHit}{See ChAMP manual.}

\item{filterXY}{See ChAMP manual.}

\item{force}{See ChAMP manual.}

\item{arraytype}{See ChAMP manual.}

\item{sampleSheet.csv}{An optional argument, that allows to add the name of a specific sample sheet using regular expressions.}

\item{preproc}{Choice of the preprocessing method of the rgSet. To use Noob dye bias and background correction use "Noob". The function is available in the \emph{minfi} package. Using "Raw" is the standard \emph{ChAMP} method and does not include background correction.}

\item{dyeMethod}{To date single sample procedure of Noob ("single") is the only choice here.}
}
\value{
A list including the rgSet, mSet, beta-value matrix, detection p-value matrix, and a matrix of intensity values. See \emph{ChAMP} manual for more detail.
}
\description{
This is a modification of the function "champ.load" from the \emph{ChAMP} package, which allows to use the Noob background correction. The original function from \emph{ChAMP} loads data from IDAT files to calculate intensity and produce quality control images. Furthermore, a new argument has been added that allows to load a specific sample sheet using a regular expression. This avoids the errors of the normal loading function when multiple csv files are found in the working directory. Only the new arguments are presented here. The other arguments can be found in the \emph{ChAMP} manual here: https://rdrr.io/bioc/ChAMP/man/champ.load.html.
The modification are minor, see line 94 and 104 to 111, everything else is from the \emph{ChAMP} package created by Yuan Tian \link{cre,aut}, Tiffany Morris \link{ctb}, Lee Stirling \link{ctb}, Andrew Feber \link{ctb}, Andrew Teschendorff \link{ctb}, Ankur Chakravarthy \link{ctb}. \emph{ChAMP} is licensed under GPL-3.
}
\examples{
\dontrun{myLoad_450K <- champ.load_extended(method = "minfi",
                                 arraytype = "450K",
                                 sampleSheet.csv ="^someSampleSheet.csv$",
                                 preproc = "Noob",
                                 dyeMethod = "single")}
}
