% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBMIQ.R
\name{checkBMIQ}
\alias{checkBMIQ}
\title{Plot the beta distribution of Infinium I and II probes before and after BMIQ-correction.}
\usage{
checkBMIQ(betaBefore, betaAfter, array = "450K", sampleChoice = NULL)
}
\arguments{
\item{betaBefore}{The beta-value matrix before BMIQ correction. The rows should be CpG probes and the columns samples.}

\item{betaAfter}{The beta-value matrix after BMIQ correction. The rows should be CpG probes and the columns samples.}

\item{array}{Selection of the array type. For EPIC data use "EPIC" and everything else will be interpreted as HM450 data. By default HM450 data is assumed.}

\item{sampleChoice}{To check individual samples, enter a sample name as it appears in the column names. Multiple samples can be entered with regular expressions, e.g. ""aPxxx|aPyy"".}
}
\value{
No return. A plot should appear in the \emph{Plots} pane.
}
\description{
This function can be used to evaluate the beta-value distribution of Infinium I and II probes before and after probe bias correction with BMIQ. Although implemented to check BMIQ performance, this function can be used to evaluate Infinium I and II probe distribution at any step of the pipeline and is also not limited to BMIQ-corrected beta-value matrices. By default, the function represents the mean density of Infinium I and II probes across all samples. It may be advisable to check for updates to \emph{ChAMP} and \emph{ChAMPdata} as new EPIC manifests are released from time to time.
}
\examples{
\dontrun{checkBMIQ(betaBefore = beta,
                    betaAfter = beta_BMIQcorrected,
                    array = "450K")}
}
