% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohensD.R
\name{cohensD}
\alias{cohensD}
\title{Cohen's D}
\usage{
cohensD(meanA, meanB, sdA, sdB, nA, nB, unbiased = F, verbose = F)
}
\arguments{
\item{meanA}{Mean of Group A}

\item{meanB}{Mean of Group B}

\item{sdA}{SD of Group A}

\item{sdB}{SD of Group B}

\item{nA}{Sample size of Group A}

\item{nB}{Sample size of Group B}

\item{unbiased}{If False the sample size will be assumed to be equal}

\item{verbose}{Extra printed information}
}
\value{
Cohen's D
}
\description{
This function calculates Cohen's D, an effect size measure. If sample sizes are not equal use the unbiased modifier.
}
\examples{
GroupA <- rnorm(1000,10, 2)
GroupB <- rnorm(1000,20, 4)

meanGroupA <- mean(GroupA)
meanGroupB <- mean(GroupB)

sigmaGroupA <- sd(GroupA)
sigmaGroupB <- sd(GroupB)

cohensD(meanA = meanGroupA,
        meanB = meanGroupB,
        sdA = sigmaGroupA,
        sdB = sigmaGroupB,
        nA = 1000, nB = 1000,
        unbiased = TRUE,
        verbose = TRUE)
}
