% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohensD_Paired.R
\name{cohensD_Paired}
\alias{cohensD_Paired}
\title{Cohen's D for paired data}
\usage{
cohensD_Paired(meanD, sigmaD)
}
\arguments{
\item{meanD}{Mean of the differences between measurements}

\item{sigmaD}{SD of the differences between measurements}
}
\value{
Cohen's D for paired data. Adapted from \emph{rstatix}
}
\description{
This function calculates Cohen's D for paired data. The formula was taken from the \emph{rstatix} package.
}
\examples{
before <- rnorm(1000, 10, 2)
after <- rnorm(1000,20, 4)
D <- before - after
meanD <- mean(D)
sigmaD <- sd(D)
cohensD_Paired(meanD = meanD,
               sigmaD = sigmaD)
}
