% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litl.R
\name{litl}
\alias{litl}
\title{\code{litl()}}
\usage{
litl(x, y, z = NULL, ntiles_x = 10, ntiles_z = NULL, plot_line = "lm")
}
\arguments{
\item{x}{A double, integer, or logical predictor variable.}

\item{y}{A logical (\code{TRUE/FALSE/NA}) or binary numeric (\code{1/0/NA})
outcome variable.}

\item{z}{An optional second predictor variable, handled variously depending
on its type (see \bold{Details}).}

\item{ntiles_x}{A double or integer value indicating how many buckets to
divide \code{x} into (10 by default).}

\item{ntiles_z}{An optional double or integer value indicating how many
buckets to divide \code{z} into (if \code{z}) is provided.}

\item{plot_line}{The type of line to be fit in the linearity-assessment plot:
\code{'lm'} (default), \code{'loess'}, or \code{NULL} (no line).}
}
\value{
Invisibly, a two-element list: The first element, \code{[['data']]},
contains the data displayed in the plot; the second, \code{[['plot']]}, is
the plot object, with class \code{ggplot}.

\code{x}, \code{y}, and \code{z} (if given) must be the same length.
\code{NA} values are allowed, although they are removed before the
calculative work proceeds.

If any of the buckets of \code{x} (optionally stratified by \code{z}) have
< 20 cases, a warning is thrown. If P(Y=1) is 1 or 0 in any bucket(s),
rendering log(P/(1-P)) undefined, a note indicating so (and indicating where)
appears on the plot.
}
\description{
\code{litl()} = \code{l}inear \code{i}n \code{t}he \code{l}ogit.
\code{litl()} generates plots for graphically assessing the reasonableness
of the linearity assumption for one- and two-variable logistic regression
models under consideration.
}
\details{
\code{litl()} assesses linearity in the logit between one or two
predictor variables, \code{x} and \code{z}, and a binary outcome, \code{y}.
\code{x} and \code{y} are required: \code{litl()} divides \code{x} into
\code{ntiles_x} quantile buckets (10 by default), calculates the log-odds
of \code{y} in each bucket, and generates a corresponding plot, as described
Eventual-StatLab-Link and in Harrell (2015, \emph{Regression Modeling Strategies}).
If \code{z} is given, observations are stratified before quantile groups are
formed based on \code{x}:

\itemize{
 \item{If \code{z} is numeric and not binary, and if a numeric value of
 \code{ntiles_z} is given, observations are stratified into \code{ntiles_z}
 quantile groups based on \code{z}, and observations within groups are then
 bucketed based on \code{x}.}
 \item{If \code{z} is not numeric and/or if \code{z} is binary, observations
 are stratified based on unique values of \code{z} before they are bucketed
 based on \code{x}.}}

 By default, a least-squares line(s) of best fit is added to the resulting
 plot: If no \code{z} is provided, one line is added across buckets of
 \code{x}; if \code{z} is provided, a line is added for each strata of
 \code{z}. A LOESS line can be added instead by setting \code{plot_line =
 'loess'}, and the line(s) can be omitted by setting \code{plot_line = NULL}.
}
\examples{
set.seed(11)
xx <- rnorm(n = 2500, mean = 1, sd = 1)
zz <- rbinom(2500, 1, .5)
logit_y <- .5*xx + 1*zz + .5*xx*zz
prob_y <- exp(logit_y) / (1 + exp(logit_y))
yy <- rbinom(length(prob_y), 1, prob_y)
yy <- ifelse(yy == 1, TRUE, FALSE)
litl(x = xx, y = yy, z = zz)

}
