% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCalcPlot.R
\name{powerCalcPlot}
\alias{powerCalcPlot}
\title{PLotting PowerCalc Results}
\usage{
powerCalcPlot(
  ...,
  sampleSizeSteps = c(5, 10, 15, 20, 25, 30, 35, 40, 45, 50),
  returnCleanPlotObj = FALSE,
  explicitNames = NULL
)
}
\arguments{
\item{...}{Any data.frames produced by powerCalc}

\item{sampleSizeSteps}{Estimated sample sizes for which the number of CpG-probes with power of 0.8 are calculated. Represents the range that will be plotted. By default this is "c(5,10,15,20,25,30,35,40,45,50)".}

\item{returnCleanPlotObj}{Returns an empty ggplot2 object that can be modified. Is set to "FALSE" by default.}

\item{explicitNames}{The explicit names of the data.frames inputted in the dots argument. Maybe useful for plots shown in presentations. Assumes same order as the data.frames. Is not used by default.}
}
\value{
A plot or a ggplot object depending on returnCleanPlotObj choice.
}
\description{
PLotting PowerCalc Results
}
